requirements.txt
google-generativeai
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
googlemaps
rich
python-dotenv